@@ -1,12 +1,14 @@
-require "test_helper"
+# frozen_string_literal: true
+
+require 'test_helper'
 
 class VoiceTest < Test::Unit::TestCase
   include ESpeak
 
   def test_all
     all_voices = Voice.all
-    assert all_voices.size > 0
-    assert %w(M F).include?(all_voices.first.gender)
+    assert all_voices.size.positive?
+    assert %w[M F].include?(all_voices.first.gender)
   end
 
   def test_find_by_language
